Refactored DOI update logic to a util
